GH-134 Split converter interface into Serializer and Deserializer
- Allow plugins which can only read or write
- Most important example: plugin for old rosbags
