+++ rosbag2/include/rosbag2/converter.hpp
@@ -23,7 +23,7 @@
 #include "rosbag2/converter_options.hpp"
 #include "rosbag2/serialization_format_converter_factory.hpp"
 #include "rosbag2/serialization_format_converter_factory_interface.hpp"
-#include "rosbag2/serialization_format_converter_interface.hpp"
+#include "rosbag2/converter_interfaces/serialization_format_converter_interface.hpp"
 #include "rosbag2/types.hpp"
 #include "rosbag2/visibility_control.hpp"
