+++ rosbag2/include/rosbag2/serialization_format_converter_factory_interface.hpp
@@ -18,7 +18,7 @@
 #include <memory>
 #include <string>
 
-#include "rosbag2/serialization_format_converter_interface.hpp"
+#include "rosbag2/converter_interfaces/serialization_format_converter_interface.hpp"
 #include "rosbag2/visibility_control.hpp"
 
 namespace rosbag2
