+++ rosbag2/test/rosbag2/converter_test_plugin.hpp
@@ -19,7 +19,7 @@
 #include <memory>
 #include <string>
 
-#include "rosbag2/serialization_format_converter_interface.hpp"
+#include "rosbag2/converter_interfaces/serialization_format_converter_interface.hpp"
 
 class ConverterTestPlugin : public rosbag2::SerializationFormatConverterInterface
 {
